Use cipher.NewGCMWithRandomNonce in Golang 1.24+
Use GCM in a way that allows the Go cryptographic module to assume the
repsonsibiltiy for verifying that Nonces are generated in a way that
complies with FIPS140-3 Annex C.H. Section 2 so that the implementation
is compatible with the FIPS cryptographic module in Go 1.24+ when
running in FIPS-only mode, i.e., GODEBUG=fips140=only.

Use a distinct implementation for Go 1.24+ and Go <1.24 using a build
tag pair in order to continue supporting older Go versions prior to the
Go 1.24 introduction of crypto/cipher.NewGCMWithRandomNonce.