Add testutil functions for checking if FIPS is enabled
After finding that the pattern of needing the check if FIPS is enabled
and selectively disable tests if it is enabled, while maintaining
compatibility with Go 1.23, was recurring often in the code, create a
public function in testutil for checking if FIPS is enabled.

Use this function to skip non-FIPS tests in testutil when operating
under FIPS mode.