Establish a FIPS baseline for streamingaead
Mask the non-FIPS-conforming use cases of streamingaead. That includes
all use of AES-GCM-HKDF because that Key/Nonce generation scheme is not
validated under any method of FIPS-140-3 Annex C.H that can be
implemented in terms of the Go cryptographic module, as well as all use
cases of AES-CTR-HMAC which use SHA1 hashes.