Skip tests that use HKDF in a non-FIPS-compliant way
HDKF in FIPS mode cannot have a salt shorter than 112 bits in FIPS mode,
which is not the case for all of the test cases. We therefore disable
those test cases if we detect that the test environment is FIPS only so
that they will continue to work in non-FIPS environments without
stopping FIPS environments where those tests are invalid from passing
tests.